Refactor the mailer injection using '@Named' annotation
